+++ redwood-treehouse-host/src/commonMain/kotlin/app/cash/redwood/treehouse/ChangeListRenderer.kt
@@ -19,6 +19,7 @@ import app.cash.redwood.protocol.EventSink
 import app.cash.redwood.protocol.SnapshotChangeList
 import app.cash.redwood.protocol.host.ProtocolBridge
 import app.cash.redwood.protocol.host.ProtocolMismatchHandler
+import app.cash.redwood.protocol.host.hostRedwoodVersion
 import kotlinx.serialization.json.Json
 
 /**
@@ -40,6 +41,8 @@ public class ChangeListRenderer<W : Any>(
   ) {
     view.reset()
     val bridge = ProtocolBridge(
+      // Use latest host version as the guest version to avoid any compatibility behavior.
+      guestVersion = hostRedwoodVersion,
       container = view.children,
       factory = view.widgetSystem.widgetFactory(
         json,